Extracted pulsing from NN into Phenotype
Also removed unnecessary serializer, which hampers develpment of new code. These should only be reimplemented once no major changes are expected, as they (are supposed to) help performance only (with some exceptions, like Phenotype, which has to call a method to add event listeners, something which is still not implemented generically in LEX).